refactor: class sanitizer new file
