+++ EMS/helpers/src/Html/HtmlSanitizerConfigBuilder.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace EMS\Helpers\Html;
 
+use EMS\Helpers\Tests\Unit\Html\HtmlClassSanitizer;
 use Symfony\Component\HtmlSanitizer\HtmlSanitizerConfig;
-use Symfony\Component\HtmlSanitizer\Visitor\AttributeSanitizer\AttributeSanitizerInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
 class HtmlSanitizerConfigBuilder
@@ -52,7 +52,7 @@ class HtmlSanitizerConfigBuilder
     public function build(): HtmlSanitizerConfig
     {
         $config = new HtmlSanitizerConfig();
-        $config = $config->withAttributeSanitizer($this->classSanitizer());
+        $config = $config->withAttributeSanitizer(new HtmlClassSanitizer($this->classes));
 
         foreach ($this->configSettings as $setting => $value) {
             $config = match ($setting) {
@@ -115,45 +115,5 @@ class HtmlSanitizerConfigBuilder
         ;
 
         return $optionsResolver;
-    }
-
-    private function classSanitizer(): AttributeSanitizerInterface
-    {
-        return new class($this->classes) implements AttributeSanitizerInterface {
-            /** @param array<mixed>|array{ allow: string[], drop: string[], replace: string[]} $settings */
-            public function __construct(private readonly array $settings = [])
-            {
-            }
-
-            public function getSupportedElements(): ?array
-            {
-                return null;
-            }
-
-            public function getSupportedAttributes(): ?array
-            {
-                return ['class'];
-            }
-
-            public function sanitizeAttribute(string $element, string $attribute, string $value, HtmlSanitizerConfig $config): ?string
-            {
-                $classes = \explode(' ', $value);
-                $classNames = \array_filter($classes, 'trim');
-
-                if (\count($this->settings['allow']) > 0) {
-                    $classNames = \array_filter($classNames, fn (string $className) => \in_array($className, $this->settings['allow']));
-                }
-
-                if (\count($this->settings['drop']) > 0) {
-                    $classNames = \array_filter($classNames, fn (string $className) => !\in_array($className, $this->settings['drop']));
-                }
-
-                if (\count($this->settings['replace']) > 0) {
-                    $classNames = \array_map(fn (string $className) => $this->settings['replace'][$className] ?? $className, $classNames);
-                }
-
-                return \count($classNames) > 0 ? \implode(' ', $classNames) : null;
-            }
-        };
     }
 }
